+++ src/spec/lists.spec.ts
@@ -29,6 +29,7 @@ describe('lists expressions', () => {
       expect(result).toStrictEqual([1, 2, 3])
     })
 
+    
     it.todo('should throw an error if lists have different types', () => {
       const expr = '[1, true]'
 
@@ -120,131 +121,15 @@ describe('lists expressions', () => {
       expect(result).toStrictEqual([1, 2])
     })
 
+    // Shall we throw an error if lists have different types?
+    // The original implementation does that if we put literals
+    // but no in case on context. So for now we will not throw an error
     it.todo('should throw an error if lists have different types', () => {
-      const expr = '[1, 2] + [true, false]'
+      const expr = '[1] + [true]'
 
       const result = () => evaluate(expr)
 
       expect(result).toThrow(new CelTypeError(Operations.logicalAnd, true, 1))
-    })
-  })
-
-  // TODO move it to comparison.spec.ts
-  describe('in', () => {
-    it('should return false for element in empty list', () => {
-      const expr = '1 in []'
-
-      const result = evaluate(expr)
-
-      expect(result).toBe(false)
-    })
-
-    it('should return true for element the only element on the list', () => {
-      const expr = '1 in [1]'
-
-      const result = evaluate(expr)
-
-      expect(result).toBe(true)
-    })
-
-    it('should return true for element the first element of the list', () => {
-      const expr = '"first" in ["first", "second", "third"]'
-
-      const result = evaluate(expr)
-
-      expect(result).toBe(true)
-    })
-
-    it('should return true for element a middle element of the list', () => {
-      const expr = '3 in [5, 4, 3, 2, 1]'
-
-      const result = evaluate(expr)
-
-      expect(result).toBe(true)
-    })
-
-    it('should return true for element the last element of the list', () => {
-      const expr = '3 in [1, 2, 3]'
-
-      const result = evaluate(expr)
-
-      expect(result).toBe(true)
-    })
-
-    it('should return false for element not in the list', () => {
-      const expr = '3 in [1, 2]'
-
-      const result = evaluate(expr)
-
-      expect(result).toBe(false)
-    })
-
-    it.todo(
-      'should thrown an error if used on something else than list',
-      () => {
-        const expr = '"a" in "asd"'
-
-        const result = () => evaluate(expr)
-
-        // TODO fix type
-        expect(result).toThrow(new CelTypeError(Operations.in, 123, 123))
-      }
-    )
-  })
-
-  // TODO create a separate file for this
-  describe('size', () => {
-    describe('list', () => {
-      it('should return 0 for empty list', () => {
-        const expr = 'size([])'
-
-        const result = evaluate(expr)
-
-        expect(result).toBe(0)
-      })
-
-      it('should return 1 for one element list', () => {
-        const expr = 'size([1])'
-
-        const result = evaluate(expr)
-
-        expect(result).toBe(1)
-      })
-
-      it('should return 3 for three element list', () => {
-        const expr = 'size([1, 2, 3])'
-
-        const result = evaluate(expr)
-
-        expect(result).toBe(3)
-      })
-    })
-
-    describe('string', () => {
-      it('should return 0 for empty string', () => {
-        const expr = 'size("")'
-
-        const result = evaluate(expr)
-
-        expect(result).toBe(0)
-      })
-
-      it('should return length of string', () => {
-        const expr = 'size("abc")'
-
-        const result = evaluate(expr)
-
-        expect(result).toBe(3)
-      })
-    })
-
-    it.todo('should thrown an error if operator is not string or list', () => {
-      const expr = 'size(123)'
-
-      const result = () => evaluate(expr)
-
-      // TODO fix type
-      expect(result).toThrow(new CelTypeError(Operations.addition, 123, 123))
     })
   })
 })
